Create Application class to wrap the main window
The window should be able to be destroyed and recreated as needed,
while the application itself remains persistent. Plus it didn't make
sense for the window object to be responsible for spinning up the XMPP
client. The Application object now spins up XMPP and has the functions
for producing and consuming CAPs. Eventually we'll also connect to
D-Bus and do libunique stuff from here as well.