stormee-inject initial commit
stormee-inject is for authenticated, programatic interaction with the
Stormee Google App Engine. It is only useful to the Stormee GAE
developer (me), but I'll put it in git because I need the version
control for myself.

This initial commit has the OAuth 1.0 (that's all GAE supports atm)
all working, and stores the credentials in the GNOME keyring (I'd use
libsecret but its not in Fedora 17). So, no secrets are in the source
anymore and I can finally commit :)

Also, the secrets are loaded into local memory via
gcr_secure_memory_alloc() Just for fun :)